refactor: Extract FileRegistryEntry validation methods to mixin
