+++ lib/kiba/extend/registry/file_registry_entry.rb
@@ -12,14 +12,12 @@ module Kiba
       #   itself and carry its own errors/warnings
       class FileRegistryEntry
         include EntrySummarizable
+        include EntryValidatable
 
         attr_reader :path, :key,
           :creator, :supplied, :dest_special_opts, :desc, :lookup_on, :tags,
           :message, :dest_class, :dest_opt, :src_class, :src_opt, :type,
           :valid, :errors, :warnings
-
-        # allowed types
-        TYPES = :file, :fileset, :enum, :lambda
 
         # @param key [Symbol] Full job key associated with reghash
         # @param reghash [Hash] File data. See {file:doc/file_registry_entry.md}
@@ -88,63 +86,12 @@ module Kiba
           errors[err.class.name] = err.message
         end
 
-        def path_required?
-          chk = [dest_class, src_class].map do |klass|
-            klass.requires_path?
-          end
-          return false if chk.uniq == [false]
-
-          true
-        end
-
         def validate
           validate_path
           validate_creator
           validate_type
           validate_lookup
           @valid = true if errors.empty?
-        end
-
-        def validate_creator
-          return if supplied
-          return if creator.is_a?(Kiba::Extend::Registry::Creator)
-
-          @creator = nil
-          @errors[:missing_creator_for_non_supplied_file] = nil
-        end
-
-        def validate_lookup
-          return unless lookup_on
-
-          supplied ? validate_supplied_lookup : validate_job_lookup
-        end
-
-        def validate_job_lookup
-          return if dest_class.as_source_class
-            .respond_to?(:is_lookupable)
-
-          @errors[:cannot_lookup_from_nonCSV_destination] = nil
-        end
-
-        def validate_supplied_lookup
-          return if src_class.name.end_with?("CSV")
-
-          @errors[:cannot_lookup_from_nonCSV_supplied_source] = nil
-        end
-
-        def validate_path
-          if path_required? && !path
-            @errors[:missing_path] = nil
-            return
-          end
-
-          @path = Pathname.new(path) if path
-        end
-
-        def validate_type
-          return if TYPES.any?(@type)
-
-          @errors[:unknown_type] = @type
         end
       end
     end
